fix(fanout): a repeat split round is asked a different question
The deep-researcher run that started this finished its fan-out, ran
analyze, routed back through gather, and re-entered the same fan-out
stage. The split prompt was byte for byte the one it had already
answered, while conversation still carried the first split, the workers'
consolidated report and the analysis built on top of it. The model read
all that and answered "I have completed the research", which was true and
was not a list of work items.

frame_split_round fires on StageJustEntered, so it sees every path into
the stage, and says the three things the second split needed to know:
which round this is, which items already came back, and that calling
submit_work_items with an empty items array is a real answer. That last
one is the load-bearing part. Parking with no work items already flowed
straight to the merge stage; nothing told the model it could say so, and
the only other way to say "the work is done" was prose.

A first entry gets none of it - no framing, no extra tokens, no change
for a run that splits once. The ids come from PreviousWorkItems, recorded
by each successful split; a restart drops the component, and the framing
still fires without the list, because "you have been here before" is the
part that matters.

The parsing moved to fanout/split_parse.rs, with its tests, to stay under
the file-size limit.
