+++ crates/leviath-runtime/src/world.rs
@@ -513,7 +513,17 @@ impl PipelineWorld {
                 // the stage just entered, holding `ReadyToInfer` until the
                 // answers land - so the stage's first request carries the fresh
                 // values rather than the previous stage's.
-                (run_stage_enter_hooks, crate::stage_seeds::start_stage_seeds).chain(),
+                //
+                // `frame_split_round` rides along: it tells a re-entered fan-out
+                // stage it has been here before, and wants the same window - set
+                // last, so the framing is the final thing in front of the
+                // stage's first request.
+                (
+                    run_stage_enter_hooks,
+                    crate::stage_seeds::start_stage_seeds,
+                    crate::fanout::frame_split_round,
+                )
+                    .chain(),
                 sync_tool_stages,
                 // Store any finished run title, then start newly-marked ones.
                 // Collect precedes persistence so a landed title is written on
